lagd en ny klasse Tasks hvor alle tidligere klasser som var tasks skal inn. Har også begynt å integrere disse inn i views
