+++ pig/templates/show_divisions.html
@@ -24,8 +24,8 @@ PURPOSE: Provide a page where users can view all the divisions they are signed u
                 {%  for division in divisions_created %}
                 <tr>
                     <td>{{ division.name }}</td>
-                    <td><button onclick="copyToClipboard('{{ ta_links[loop.index0] }}')">Kopier lenken</button></td>
-                    <td><button onclick="copyToClipboard('{{ student_links[loop.index0] }}')">Kopier lenken</button></td>
+                    <td><button onclick="copyToClipboard('{{ leader_links[loop.index0] }}')">Kopier lenken</button></td>
+                    <td><button onclick="copyToClipboard('{{ member_links[loop.index0] }}')">Kopier lenken</button></td>
                 </tr>
                 {% endfor %}
             </table>
