+++ pig/views.py
@@ -7,11 +7,11 @@ from pig.login.login_handler import LoginHandler
 from pig.login.registration_handler import RegistrationHandler
 from pig.scripts.create_division import Task_CreateDivision
 import pig.scripts.encryption as encryption
-from pig.scripts.get_divisions import Task_GetDivisions
 from pig.scripts.register_user import Task_RegisterUser
 from pig.scripts.Task_GetDivisionsWhereLeader import Task_GetDivisionsWhereLeader
 from pig.db.database import Database
 from pig.scripts.DbGetters import DbGetters
+from pig.scripts.Tasks import Tasks
 
 import sys
 
@@ -24,7 +24,6 @@ database = Database(app)
 from pig.db.models import *
 
 
-
 pig_key = "supersecretpigkey"
 app.secret_key = pig_key
 login_manager = LoginManager()
@@ -34,12 +33,12 @@ login_handler, registration_handler = LoginHandler(database, User), Registration
 
 
 division_creator = Task_CreateDivision(database, Division, Parameter, NumberParam, EnumVariant)
-get_divisions = Task_GetDivisions(database, User)
 get_divisions_where_leader = Task_GetDivisionsWhereLeader(database, Division, user_division)
 division_registrator = Task_RegisterUser(database, User, Division, user_division)
 db_getters = DbGetters(
                 database, User, Division, Group, Parameter, Value, NumberParam, EnumVariant,
                 user_division, user_group, division_parameter, parameter_value, user_division_parameter_value)
+tasks = Tasks()
 
 database.get_session().commit()
 #This code is being used by the login_manager to grab users based on their IDs. Thats how we identify which user we
@@ -129,13 +128,11 @@ def home():
 @app.route("/show_divisions")
 @login_required
 def show_divisions():
-    divisions_participating2 = db_getters.get_all_divisions_where_member_or_leader_for_given_user(current_user=current_user)
-    divisions_created2 = db_getters.get_all_divisions_where_creator_for_given_user(current_user=current_user)
-    divisions_participating, divisions_created, ta_links, student_links = get_divisions.fetch_divisions(current_user, pig_key)
-    print(divisions_participating == divisions_participating2)
-    print(divisions_created == divisions_created2)
+    divisions_participating = db_getters.get_all_divisions_where_member_or_leader_for_given_user(current_user=current_user)
+    divisions_created = db_getters.get_all_divisions_where_creator_for_given_user(current_user=current_user)
+    leader_links, member_links = tasks.generate_links(pig_key,divisions_created)
     return render_template("show_divisions.html", user=current_user,
-                           divisions_participating=divisions_participating, divisions_created=divisions_created, ta_links=ta_links, student_links=student_links)
+                           divisions_participating=divisions_participating, divisions_created=divisions_created, leader_links=leader_links, member_links=member_links)
 
 @app.route("/logout")
 @login_required
